Don't show "Accept both" when not both can be accepted. Don't show "Swap" when it does not have an effect.
